requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions>=1.12.0
fastapi
pydantic>=1.2.0,<2.0.0
uvicorn>=0.15.0,<0.16.0
requests==2.26.0
pytest==6.2.5